# perf(python): order(s) of magnitude speedup when initialising `List` dtype `Series` from 2D numpy array