Closes #12669.

Instead of using a naïve conversion inside a Python loop, we now reshape the input numpy array as 1D, fast-initialise a Series from that, and then defer to a Rust-side reshape operation to optimally recover the given dimensions.

In the specific example given in the issue this results in initialising the frame ~2,500x faster  🚀
(The degree of speedup depends on the dimensionality, but it's typically ~1-3 orders of magnitude).

### Examples

* **Average case**

  ```python
  from codetiming import Timer
  import polars as pl
  import numpy as np

  a = np.arange(1_000_000).reshape(-1, 5)
  with Timer():
      df = pl.DataFrame({"a": a})

  # shape: (200_000, 1)
  # ┌────────────────────────────┐
  # │ a                          │
  # │ ---                        │
  # │ list[i64]                  │
  # ╞════════════════════════════╡
  # │ [0, 1, … 4]                │
  # │ [5, 6, … 9]                │
  # │ …                          │
  # │ [999990, 999991, … 999994] │
  # │ [999995, 999996, … 999999] │
  # └────────────────────────────┘
  ```
  ```
  Before: 0.6246 seconds
   After: 0.0151 seconds
  -------------------------
  Speedup: ~41x 
  -------------------------
  ```

* **Extreme case**
  ```python
  a = np.arange(100_000_000).reshape(-1, 1)
  with Timer():
      df = pl.DataFrame({"a": a})

  # shape: (100_000_000, 1)
  # ┌────────────┐
  # │ a          │
  # │ ---        │
  # │ list[i64]  │
  # ╞════════════╡
  # │ [0]        │
  # │ [1]        │
  # │ …          │
  # │ [99999998] │
  # │ [99999999] │
  # └────────────┘
  ```
  ```
  Before: 349.3047 seconds
   After:   0.0738 seconds
  -------------------------
  Speedup: ~4,700x 
  -------------------------
  ```

---

### Also

* Splits up the overly-long `test_init_ndarray` func into a few more manageable pieces.
* Fixes an edge-case (ref: #10725) when initialising square numpy arrays; can resolve an otherwise ambiguous tiebreak by checking the flags to confirm c-contiguous vs fortran row/col major order.