@alexander-beedie 2023-11-24
(Will fix the test dtype in the morning! ;)