@alexander-beedie 2023-11-25
> Only ~2,500x faster?

I'll try harder 🙏